app/src/main/python/text_normalizer/requirements.txt
pyyaml
flask
waitress
pandas
python-logstash-async
dlinfo
Bio